+++ src/pl02/pl02.cpp
@@ -167,7 +167,7 @@ cPlAda::cPlAda()
     EspDataLoad((u32) PL_ARC_PTR(arc, 0x1A), 3, 0);
     startUp();
     if (pG->pl_costume == 1) {
-        EstSet((int) this, -1, 0, 0, 0, 0x59, 0x800, 0, 0, 0);
+        EstSet(this, -1, 0, 0, 0, 0x59, 0x800, 0, 0, 0);
     }
     pFootShadowTbl = pl_fs_tbl;
 }
